hot-fix/id type change
- id type 을 int -> string 으로 변경
- patch dialog 실행

+++ app/src/main/java/com/meongbyeol/taxishare/view/main/taxisharelist/TaxiShareListAdapter.kt
@@ -247,7 +247,7 @@ class TaxiShareListAdapter :
                         R.drawable.background_already_participate_color,
                         R.color.light_gray
                     )
-                } else if (Constant.CURRENT_USER.studentId == uid.toInt()) {
+                } else if (Constant.CURRENT_USER.studentId == uid) {
                     changeButtonState(
                         view,
                         String.format(view.resources.getString(R.string.my_taxi_share_title)),
